// frontend/src/models/User.ts
/**
 * Interfaz que representa un usuario del sistema.
 * Contiene los datos básicos del usuario.
 */
export type User = {
    /** ID del usuario */
    id?: number

    /** Nickname del usuario */
    nickname?: string

    /** Nombre real del usuario */
    name?: string

    /** Apellidos del usuario */
    lastname?: string

    /** Correo electrónico del usuario */
    email?: string

    /** Fecha de nacimiento en formato ISO */
    birthdate?: string

    /** URL de la imagen de perfil */
    profile_url?: string
}